testing/testy-tester/requirements.txt
adal==1.2.0
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
bs4==0.0.0
cachetools==3.0.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
fabric==2.3.1
google-auth==1.6.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
invoke==1.1.1
Jinja2==2.10
junit-xml==1.8
kubernetes==7.0.0
lxml==4.2.4
MarkupSafe==1.0
oauthlib==2.1.0
paramiko==2.4.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyHamcrest==1.9.0
PyJWT==1.7.0
pymongo==3.7.1
PyNaCl==1.2.1
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pyvmomi==6.7.0.2018.9
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
selenium==3.14.1
service-identity==17.0.0
simplejson==3.16.0
six==1.11.0
suds-jurko==0.6
tabulate==0.8.2
Twisted==18.7.0
urllib3==1.23
websocket-client==0.54.0
Werkzeug==0.14.1
zope.interface==4.5.0

# VMWare Deps
nsx-policy-python-sdk
vapi-client-bindings
vapi-common-client
vapi-runtime
vmc-app-python-sdk
vmc-client-bindings
